// src/shared/model/users.model.ts
export interface IUsers{
    id: number;
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    car_model_year: number;
    nationality: string;
}
// { id: number; first_name: string; last_name: string; email: string; gender: string; car_model_year: number; nationality: string; }

export interface DataImage{
    image: string,
    name: string,
    id: number
}